split multiple auto-config classes into separate files
