+++ spring-grpc-spring-boot-autoconfigure/src/main/java/org/springframework/grpc/autoconfigure/server/security/GrpcNativeSecurityConfigurerAutoConfiguration.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright 2024-2024 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.springframework.grpc.autoconfigure.server.security;
+
+import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Conditional;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.grpc.autoconfigure.server.GrpcServerFactoryAutoConfiguration;
+import org.springframework.grpc.server.security.GrpcSecurity;
+import org.springframework.security.config.ObjectPostProcessor;
+import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
+import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
+
+@ConditionalOnBean(ObjectPostProcessor.class)
+@Configuration(proxyBeanMethods = false)
+@Conditional(GrpcServerFactoryAutoConfiguration.OnNativeGrpcServerCondition.class)
+public class GrpcNativeSecurityConfigurerAutoConfiguration {
+
+	@Bean
+	public GrpcSecurity grpcSecurity(ObjectPostProcessor<Object> objectPostProcessor,
+			AuthenticationConfiguration authenticationConfiguration, ApplicationContext context) throws Exception {
+		AuthenticationManagerBuilder authenticationManagerBuilder = authenticationConfiguration
+			.authenticationManagerBuilder(objectPostProcessor, context);
+		authenticationManagerBuilder
+			.parentAuthenticationManager(authenticationConfiguration.getAuthenticationManager());
+		return new GrpcSecurity(objectPostProcessor, authenticationManagerBuilder, context);
+	}
+
+}
